backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.3.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Database Drivers for Secure Data Connector
asyncpg==0.29.0
aiomysql==0.2.0
httpx==0.25.2

# Enhanced PII Detection
presidio-analyzer
presidio-anonymizer
spacy

# Bias & Fairness Detection
fairlearn
aif360
witwidget

# Adversarial Detection
textattack
adversarial-robustness-toolbox

# Additional NLP & ML
numpy
pandas
scikit-learn
transformers
torch
